Aug 21:  Brainstorming
Today I mainly brainstormed ideas since I could not come up with one on
Friday.

The first idea might be a bit complicated. It would involve a ball of
dough flying through space flying around planets and avoiding meteors
to get to the sun to bake itself. It might be out of reach as it would
require me to create complex gravity.

The second idea is a platformer were the player controls two figures
simultaneously. One character might have an obstacle that the other one
would not have to deal with, or might have to deal with a different
obstacle concurrently.